+++ src/Http/FormRequest.php
@@ -5,6 +5,7 @@ namespace Somnambulist\Bundles\FormRequestBundle\Http;
 use BadMethodCallException;
 use InvalidArgumentException;
 use Somnambulist\Bundles\FormRequestBundle\Http\Behaviours\GetNullOrValue;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\FileBag;
 use Symfony\Component\HttpFoundation\HeaderBag;
 use Symfony\Component\HttpFoundation\InputBag;
@@ -12,10 +13,8 @@ use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\ServerBag;
 use Symfony\Component\HttpFoundation\Session\Session;
-use Symfony\Component\Security\Core\Security;
 use function array_replace_recursive;
 use function in_array;
-use function method_exists;
 use function Somnambulist\Bundles\FormRequestBundle\Resources\arrayGet;
 use function Somnambulist\Bundles\FormRequestBundle\Resources\arrayHas;
 use function Somnambulist\Bundles\FormRequestBundle\Resources\forget;
@@ -97,12 +96,7 @@ abstract class FormRequest
             return $this->source->getContent();
         }
         if ('payload' === $name) {
-            // SF 6.3
-            if (method_exists($this->source, 'getPayload')) {
-                return $this->source->getPayload();
-            }
-
-            return $this->source->request;
+            return $this->source->getPayload();
         }
 
         return null;