+++ tests/Http/FormRequestTest.php
@@ -27,7 +27,6 @@ class FormRequestTest extends TestCase
         $this->assertSame($r->attributes, $form->attributes);
         $this->assertSame($r->files, $form->files);
         $this->assertSame($r->server, $form->server);
-        $this->assertSame($r->request, $form->payload);
 
         $this->assertSame($r->query, $form->query());
         $this->assertSame($r->request, $form->request());
@@ -35,11 +34,7 @@ class FormRequestTest extends TestCase
         $this->assertSame($r->files, $form->files());
         $this->assertSame($r->server, $form->server());
 
-        if (method_exists($r, 'getPayload')) {
-            $this->assertSame($r->getPayload(), $form->payload());
-        } else {
-            $this->assertSame($r->request, $form->payload());
-        }
+        $this->assertEquals($r->getPayload(), $form->payload());
     }
 
     public function testMagicContentPassThrough()